fix bugs

+++ src/pages/Authentication/SignIn.tsx
@@ -5,13 +5,20 @@ import LogoDark from '../../images/logo/logo-dark.svg';
 import Logo from '../../images/logo/logo.svg';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-import { URL } from "../../types/constant";
+import Swal from 'sweetalert2/dist/sweetalert2.js';
+import 'sweetalert2/src/sweetalert2.scss';
+import { URL } from '../../types/constant';
 
 const SignIn: React.FC = () => {
   const [walletAddress, setWalletAddress] = useState('');
   const [password, setPassword] = useState('');
   const handleSignin = () => {
-    if (walletAddress === undefined || walletAddress === '' || password === undefined || password === '') {
+    if (
+      walletAddress === undefined ||
+      walletAddress === '' ||
+      password === undefined ||
+      password === ''
+    ) {
       toast.error('Username and password cant be null!', {
         position: 'top-right',
         autoClose: 1500,
@@ -37,13 +44,23 @@ const SignIn: React.FC = () => {
 
     Axios.request(config)
       .then((response) => {
-        toast.success('Sign in success!', {
-          position: 'top-right',
-          autoClose: 1500,
-          onClick: () => {
-            localStorage.setItem('access_token', response.data.access_token);
-            window.location.href = '/';
-          },
+        // toast.success('Sign in success!', {
+        //   position: 'top-right',
+        //   autoClose: 1500,
+        //   onClick: () => {
+        //     localStorage.setItem('access_token', response.data.access_token);
+        //     window.location.href = '/';
+        //   },
+        // });
+        Swal.fire({
+          position: 'top-end',
+          icon: 'success',
+          title: 'Sign in success',
+          showConfirmButton: false,
+          timer: 2000,
+        }).then(() => {
+          localStorage.setItem('access_token', response.data.access_token);
+          window.location.href = "/";
         });
       })
       .catch((error) => {
@@ -239,51 +256,51 @@ const SignIn: React.FC = () => {
                 </div>
 
                 <label className="mb-2.5 block font-medium text-black dark:text-white">
-                    Password
-                  </label>
-                  <div className="relative">
-                    <input
-                      type="password"
-                      value={password}
-                      onChange={(e) => {
-                        setPassword(e.target.value);
-                      }}
-                      placeholder="Enter your password"
-                      className="w-full rounded-lg border border-stroke bg-transparent py-4 pl-6 pr-10 text-black outline-none focus:border-primary focus-visible:shadow-none dark:border-form-strokedark dark:bg-form-input dark:text-white dark:focus:border-primary"
-                    />
+                  Password
+                </label>
+                <div className="relative">
+                  <input
+                    type="password"
+                    value={password}
+                    onChange={(e) => {
+                      setPassword(e.target.value);
+                    }}
+                    placeholder="Enter your password"
+                    className="w-full rounded-lg border border-stroke bg-transparent py-4 pl-6 pr-10 text-black outline-none focus:border-primary focus-visible:shadow-none dark:border-form-strokedark dark:bg-form-input dark:text-white dark:focus:border-primary"
+                  />
 
-                    <span className="absolute right-4 top-4">
-                      <svg
-                        className="fill-current"
-                        width="22"
-                        height="22"
-                        viewBox="0 0 22 22"
-                        fill="none"
-                        xmlns="http://www.w3.org/2000/svg"
-                      >
-                        <g opacity="0.5">
-                          <path
-                            d="M19.2516 3.30005H2.75156C1.58281 3.30005 0.585938 4.26255 0.585938 5.46567V16.6032C0.585938 17.7719 1.54844 18.7688 2.75156 18.7688H19.2516C20.4203 18.7688 21.4172 17.8063 21.4172 16.6032V5.4313C21.4172 4.26255 20.4203 3.30005 19.2516 3.30005ZM19.2516 4.84692C19.2859 4.84692 19.3203 4.84692 19.3547 4.84692L11.0016 10.2094L2.64844 4.84692C2.68281 4.84692 2.71719 4.84692 2.75156 4.84692H19.2516ZM19.2516 17.1532H2.75156C2.40781 17.1532 2.13281 16.8782 2.13281 16.5344V6.35942L10.1766 11.5157C10.4172 11.6875 10.6922 11.7563 10.9672 11.7563C11.2422 11.7563 11.5172 11.6875 11.7578 11.5157L19.8016 6.35942V16.5688C19.8703 16.9125 19.5953 17.1532 19.2516 17.1532Z"
-                            fill=""
-                          />
-                        </g>
-                      </svg>
-                    </span>
-                  </div>
+                  <span className="absolute right-4 top-4">
+                    <svg
+                      className="fill-current"
+                      width="22"
+                      height="22"
+                      viewBox="0 0 22 22"
+                      fill="none"
+                      xmlns="http://www.w3.org/2000/svg"
+                    >
+                      <g opacity="0.5">
+                        <path
+                          d="M19.2516 3.30005H2.75156C1.58281 3.30005 0.585938 4.26255 0.585938 5.46567V16.6032C0.585938 17.7719 1.54844 18.7688 2.75156 18.7688H19.2516C20.4203 18.7688 21.4172 17.8063 21.4172 16.6032V5.4313C21.4172 4.26255 20.4203 3.30005 19.2516 3.30005ZM19.2516 4.84692C19.2859 4.84692 19.3203 4.84692 19.3547 4.84692L11.0016 10.2094L2.64844 4.84692C2.68281 4.84692 2.71719 4.84692 2.75156 4.84692H19.2516ZM19.2516 17.1532H2.75156C2.40781 17.1532 2.13281 16.8782 2.13281 16.5344V6.35942L10.1766 11.5157C10.4172 11.6875 10.6922 11.7563 10.9672 11.7563C11.2422 11.7563 11.5172 11.6875 11.7578 11.5157L19.8016 6.35942V16.5688C19.8703 16.9125 19.5953 17.1532 19.2516 17.1532Z"
+                          fill=""
+                        />
+                      </g>
+                    </svg>
+                  </span>
                 </div>
+              </div>
 
-                <div className="mb-5">
-                  <button
-                    onClick={handleSignin}
-                    className="w-full cursor-pointer rounded-lg border border-primary bg-primary p-4 text-white transition hover:bg-opacity-90"
-                  >
-                    Sign in
-                  </button>
-                </div>
+              <div className="mb-5">
+                <button
+                  onClick={handleSignin}
+                  className="w-full cursor-pointer rounded-lg border border-primary bg-primary p-4 text-white transition hover:bg-opacity-90"
+                >
+                  Sign in
+                </button>
               </div>
             </div>
           </div>
         </div>
+      </div>
       <ToastContainer />
     </>
   );
